rearrange markup, add data
