+++ src/scss/apps.scss
@@ -28,7 +28,7 @@ body {
 
   hathi-alert-banner {
     grid-area: banner;
-    display: none;
+    // display: none;
   }
 
   hathi-website-footer {
